+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-* no relevant changes
+### Added
+* Added new 'fasp' and 'fasp amber and green' case / patient categoryfilters (#78)
 
 ## 5.6.0 / 2021-03-16
 ### Added